RRF: Stay at home after homing
Depending on the installed position of the X and Y endstops, X0, Y0 can actually be off the front corner of the bed. Further, this can cause a light collision between the BLTouch and the front left idler.

As the end user will most likely perform a homing move as the first thing they do, this will avoid any issues as much as possible.