DXF: Add all enclosure panels (imperial)
Add DXF files of all of the panels used for the full enclosure, in imperial format.

This includes a plain electronics panel which only features the required mounting holes to fasten it to the frame. The user is then free to drill their own mounting holes as they see fit.